Is it working? no it isn't. Is it ready? no it isn't. Are you sure you're ready to explore? no it ain't

+++ src/main/java/tk/knownunown/turtle/network/raknet/PacketInfo.java
@@ -6,8 +6,8 @@ package tk.knownunown.turtle.network.raknet;
 public class PacketInfo {
 
     public String placeholder = "derp";
-    public static final int ConnectedPing = 0x01;
-    public static final int UnconnectedPing = 0x02;
-    public static final int UnconnectedPingOpenConnections = 0x1c;
-    public static final int AdvertiseSystem = 0x1d;
+    public static final byte ConnectedPing = 0x01;
+    public static final byte UnconnectedPing = 0x02;
+    public static final byte UnconnectedPingOpenConnections = 0x1c;
+    public static final byte AdvertiseSystem = 0x1d;
 }
